Fix next autotaunt time calculation
Previously, the formula added one to the duration in seconds and
multiplied the result by OBJHUNT_AUTOTAUNT_DURATION_MODIFIER.  In the
extreme, if the MODIFIER is 0, the formula would not be accounting for
the taunt duration at all.

I believe the intention of the +1 second was actually to ensure that at
least one entire duration is always included, regardless of the
MODIFIER value.