Clamp next auto taunt time to a positive value
While this value is logically never negative, there are certain times
(esp. right at the beginning of a round) where its value isn't yet
trustworthy (because the server has broadcast the round update but not
set the AutoTauntDelay yet).  To avoid weird graphical glitches during
those times (or if there is a bug :P), we can just clamp this value to
something sane.