Remove the LastTauntPitch network variable
It doesn't make sense (er, to me at least) for the taunt selection UI
to use the randomly-selected pitch of the last auto-taunt.  With this
change, the UI's remembered pitch is a client-side player property that
is distinct from the last played taunt pitch.

Since this was the only use of LastTauntPitch, I have removed it.  The
pitch still affects the server's NextAutoTauntDelay calculation, but is
not needed afterwards.