Add caching to sound duration logic
Previously, sound files were opened and parsed every time a taunt is triggered to determine their
duration. Sound durations are now cached so they only need to be parsed once. The cache grows
indefinitely but this shouldn't be an issue for this gamemode since there is a fixed number of
sounds.