Remove some completely unnecessary uses of hooks
In general, I have Great Doubt in my heart about the use of hooks.  I
think it makes the code harder to read and debug.  It's easy to get
trapped in Hook Hell when hooks don't fire in the order you expect,
and it's easy to accidentally rely on hooks firing in a particular
order when GMod makes no such guarantees.

This change inlines some hooks which are invoked and attached in the
same module.  Regardless of how you feel about hooks, we can all agree
that these hooks are completely pointless.