Use model bounds instead of hitboxes for prop collisions
This is a pretty important change.  The upshot is:

  - An uncommon (but not rare!) condition where rotation-lock would set
    a nonsense hitbox is now fixed.
  - A lot of code complexity can disappear.
  - Props with scale != 1 will probably misbehave, but I haven't found
    any such props yet.

A bit more background:

For some time, I've been seeing an uncommon condition where prop
players' hitboxes would be incorrectly translated after rotation lock.
I did some (er, way too much) investigation, and learned that bone
translations can lag by up to 1 server tick.  This enables the
following fairly likely sequence of events:

  1. A player enables rotation lock.
  2. The server gets the `net` message, computes a new hitbox, and
     adjusts the player's position.
  3. player_prop_ent:Think() runs, and adjusts the prop's position to
     match the player's.
  4. The server gets a dup of the message from (2) --- or maybe the
     player just enabled/disabled rotation lock a lot.  The bone
     position still has its old value from before (2), so the server's
     computed hitbox (which depends on both the bone and entity
     positions) is computed incorrectly!

This sequence of events is disabled if either (A) step 3 does not
happen or (B) a server tick happens between 2 and 4.  However, it is
still likely enough that I was seeing it a lot.

In conclusion: bones and hitboxes are TOO COMPLICATED.  I'm fed up with
them, so I'm going to try something else.